+++ wemake_python_styleguide/visitors/ast/functions.py
@@ -6,6 +6,7 @@ from wemake_python_styleguide import constants
 from wemake_python_styleguide.logics import functions
 from wemake_python_styleguide.types import final
 from wemake_python_styleguide.violations.best_practices import (
+    BooleanPositionalArgumentViolation,
     WrongFunctionCallViolation,
 )
 from wemake_python_styleguide.visitors.base import BaseNodeVisitor
@@ -28,13 +29,25 @@ class WrongFunctionCallVisitor(BaseNodeVisitor):
                 WrongFunctionCallViolation(node, text=function_name),
             )
 
+    def _check_boolean_arguments(self, node: ast.Call) -> None:
+        for arg in node.args:
+            if isinstance(arg, ast.NameConstant):
+                # We do not check for `None` values here:
+                if arg.value is True or arg.value is False:
+                    self.add_violation(
+                        BooleanPositionalArgumentViolation(node),
+                    )
+                    break
+
     def visit_Call(self, node: ast.Call) -> None:
         """
         Used to find ``FUNCTIONS_BLACKLIST`` calls.
 
         Raises:
+            BooleanPositionalArgumentViolation
             WrongFunctionCallViolation
 
         """
         self._check_wrong_function_called(node)
+        self._check_boolean_arguments(node)
         self.generic_visit(node)